Rename Iterator Object
Make the name more obvious by renaming it from:

```
\ThriftSQL\ThriftStream()
```

to:

```
\ThriftSQL\Utils\Iterator()
```
